* commit '2db4f9968611c1c0bd8296ec0019ffd56c228ce5':
  menumanage 특수문자 공백, 숫자 예외처리
  menu price 0000 edit
  menu edit add alert
  menu delete add confirm
  img validate apply

+++ FE/src/routes/admin/AdminMenuManage/AdminMenuManagePresenter.js
@@ -83,7 +83,8 @@ const AdminMenuManagePresenter = ({
         {
           icon: <span className="gridBtn">삭제</span>,
           callback: () => {
-            deleteItem(row.id);
+            const isDelete = window.confirm(`정말로 ${row.name}을/를 삭제하시겠습니까?`);
+            isDelete && deleteItem(row.id);
           },
         },
       ],
